fix warnings.

+++ packages/rendering/source/common/ui/UIVisitors.cpp
@@ -362,7 +362,7 @@ namespace l::ui {
                     size.x *= layoutArea.mScale;
                     size.y *= layoutArea.mScale;
                     ImVec2 startPos = ImVec2(p1.x, p1.y + 0.5f * size.y);
-                    for (size_t i = 0; i < nodeValueCount - 1; i++) {
+                    for (int32_t i = 0; i < nodeValueCount - 1; i++) {
                         float xpart1 = i / static_cast<float>(nodeValueCount);
                         float xpart2 = (i+1) / static_cast<float>(nodeValueCount);
                         ImVec2 graphP1 = ImVec2(startPos.x + size.x * xpart1, startPos.y + 0.5f * nodeValues[i] * size.y);
